requirements.txt
annotated-types==0.7.0
anyio==4.3.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.2
groq==0.8.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.1
pandas==2.1.3
pathspec==0.11.1
platformdirs==3.8.0
pydantic==2.7.1
pydantic_core==2.18.2
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
telemulator3==1.4
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.0.3
Werkzeug==3.0.1
zipp==3.17.0
